refactor: add Banner component from combined Left and User Menus, updated layout and pages to use same Banner for consistency
