+++ src/lib/components/overview-page/UserMenu.svelte
@@ -14,7 +14,7 @@
     <Icon icon="tabler:menu-2" class="icon-medium" />
   </button>
 </div>
-	
+
 	
 <style>
 .user-menu {
