+++ src/routes/overview-page/+layout.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
     import { page } from "$app/state";
     import Banner from "$lib/components/overview-page/Banner.svelte";
-    import UserMenu from "$lib/components/overview-page/UserMenu.svelte";
     import Sidebar from "$lib/components/global/Sidebar.svelte";
 
     let profile_image_url = "/mock_profile_img.png";
@@ -16,22 +15,8 @@
 
 <main class="page">
     <header class="header">
-        <Banner {repo_url} {repo_path} />
-        <UserMenu username={user_name} {profile_image_url} />
+        <Banner {repo_url} {repo_path} {user_name} {profile_image_url}/>
     </header>
     {@render children()}
 </main>
-
-<Sidebar />
-
-<style>
-    .header {
-        padding-left: 2rem;
-        padding-right: 2rem;
-        padding-top: 2rem;
-        margin-bottom: 0.8125rem;
-        height: 1.375rem;
-        display: flex;
-        justify-content: space-between;
-    }
-</style>
+<Sidebar/>
