+++ static/global.css
@@ -16,6 +16,15 @@
     background-blend-mode: screen;
 }
 
+.header {
+    padding-left: 1.5rem;
+    padding-right: 1.5rem;
+    padding-top: 1.5rem;
+    margin-bottom: 0.8125rem;
+    height: 1.375rem;
+    display: flex;
+    justify-content: space-between;
+}
 /* Display fonts */
 .display-title {
     font-size: 2.625rem;
